refactor(flags): Split out functionality from models
