+++ posthog/models/feature_flag/flag_matching.py
@@ -1,265 +1,29 @@
 import hashlib
-import json
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Dict, List, Optional, Tuple, Union, cast
+from typing import Any, Dict, List, Optional, Tuple, Union
 
-from django.core.cache import cache
-from django.db import DatabaseError, models
+from django.db import DatabaseError
 from django.db.models.expressions import ExpressionWrapper, RawSQL
 from django.db.models.fields import BooleanField
 from django.db.models.query import QuerySet
-from django.db.models.signals import pre_delete
-from django.utils import timezone
-from rest_framework.exceptions import ValidationError
 from sentry_sdk.api import capture_exception
 
-from posthog.client import sync_execute
-from posthog.constants import AvailableFeature, PropertyOperatorType
-from posthog.models.cohort import Cohort
-from posthog.models.experiment import Experiment
 from posthog.models.filters import Filter
 from posthog.models.filters.mixins.utils import cached_property
 from posthog.models.group import Group
 from posthog.models.group_type_mapping import GroupTypeMapping
-from posthog.models.organization import OrganizationMembership
 from posthog.models.person import Person, PersonDistinctId
 from posthog.models.property import GroupTypeIndex, GroupTypeName
-from posthog.models.property.property import Property, PropertyGroup
-from posthog.models.signals import mutable_receiver
-from posthog.models.team.team import Team
+from posthog.models.property.property import Property
 from posthog.queries.base import match_property, properties_to_Q
 
-
-class FeatureFlag(models.Model):
-    class Meta:
-        constraints = [models.UniqueConstraint(fields=["team", "key"], name="unique key for team")]
-
-    key: models.CharField = models.CharField(max_length=400)
-    name: models.TextField = models.TextField(
-        blank=True
-    )  # contains description for the FF (field name `name` is kept for backwards-compatibility)
-
-    filters: models.JSONField = models.JSONField(default=dict)
-    rollout_percentage: models.IntegerField = models.IntegerField(null=True, blank=True)
-
-    team: models.ForeignKey = models.ForeignKey("Team", on_delete=models.CASCADE)
-    created_by: models.ForeignKey = models.ForeignKey("User", on_delete=models.CASCADE)
-    created_at: models.DateTimeField = models.DateTimeField(default=timezone.now)
-    deleted: models.BooleanField = models.BooleanField(default=False)
-    active: models.BooleanField = models.BooleanField(default=True)
-
-    rollback_conditions: models.JSONField = models.JSONField(null=True, blank=True)
-    performed_rollback: models.BooleanField = models.BooleanField(null=True, blank=True)
-
-    ensure_experience_continuity: models.BooleanField = models.BooleanField(default=False, null=True, blank=True)
-
-    def get_analytics_metadata(self) -> Dict:
-        filter_count = sum(len(condition.get("properties", [])) for condition in self.conditions)
-        variants_count = len(self.variants)
-
-        return {
-            "groups_count": len(self.conditions),
-            "has_variants": variants_count > 0,
-            "variants_count": variants_count,
-            "has_filters": filter_count > 0,
-            "has_rollout_percentage": any(condition.get("rollout_percentage") for condition in self.conditions),
-            "filter_count": filter_count,
-            "created_at": self.created_at,
-            "aggregating_by_groups": self.aggregation_group_type_index is not None,
-        }
-
-    @property
-    def conditions(self):
-        "Each feature flag can have multiple conditions to match, they are OR-ed together."
-        return self.get_filters().get("groups", []) or []
-
-    @property
-    def _payloads(self):
-        return self.get_filters().get("payloads", {}) or {}
-
-    def get_payload(self, match_val: str) -> Optional[object]:
-        return self._payloads.get(match_val, None)
-
-    @property
-    def aggregation_group_type_index(self) -> Optional[GroupTypeIndex]:
-        "If None, aggregating this feature flag by persons, otherwise by groups of given group_type_index"
-        return self.get_filters().get("aggregation_group_type_index", None)
-
-    @property
-    def variants(self):
-        # :TRICKY: .get("multivariate", {}) returns "None" if the key is explicitly set to "null" inside json filters
-        multivariate = self.get_filters().get("multivariate", None)
-        if isinstance(multivariate, dict):
-            variants = multivariate.get("variants", None)
-            if isinstance(variants, list):
-                return variants
-        return []
-
-    def get_filters(self):
-        if "groups" in self.filters:
-            return self.filters
-        else:
-            # :TRICKY: Keep this backwards compatible.
-            #   We don't want to migrate to avoid /decide endpoint downtime until this code has been deployed
-            return {
-                "groups": [
-                    {"properties": self.filters.get("properties", []), "rollout_percentage": self.rollout_percentage}
-                ],
-            }
-
-    def transform_cohort_filters_for_easy_evaluation(self):
-        """
-        Expands cohort filters into person property filters when possible.
-        This allows for easy local flag evaluation.
-        """
-        # Expansion depends on number of conditions on the flag.
-        # If flag has only the cohort condition, we get more freedom to maneuver in the cohort expansion.
-        # If flag has multiple conditions, we can only expand the cohort condition if it's a single property group.
-        # Also support only a single cohort expansion. i.e. a flag with multiple cohort conditions will not be expanded.
-        # Few more edge cases are possible here, where expansion is possible, but it doesn't seem
-        # worth it trying to catch all of these.
-
-        if len(self.cohort_ids) != 1:
-            return self.conditions
-
-        cohort_group_rollout = None
-        cohort: Optional[Cohort] = None
-
-        parsed_conditions = []
-        for condition in self.conditions:
-            cohort_condition = False
-            props = condition.get("properties", [])
-            cohort_group_rollout = condition.get("rollout_percentage")
-            for prop in props:
-                if prop.get("type") == "cohort":
-                    cohort_condition = True
-                    cohort_id = prop.get("value")
-                    if cohort_id:
-                        if len(props) > 1:
-                            # We cannot expand this cohort condition if it's not the only property in its group.
-                            return self.conditions
-                        try:
-                            cohort = Cohort.objects.get(pk=cohort_id)
-                        except Cohort.DoesNotExist:
-                            return self.conditions
-            if not cohort_condition:
-                # flag group without a cohort filter, let it be as is.
-                parsed_conditions.append(condition)
-
-        if not cohort or len(cohort.properties.flat) == 0:
-            return self.conditions
-
-        if not all(property.type == "person" for property in cohort.properties.flat):
-            return self.conditions
-
-        # all person properties, so now if we can express the cohort as feature flag groups, we'll be golden.
-
-        # If there's only one effective property group, we can always express this as feature flag groups.
-        # A single ff group, if cohort properties are AND'ed together.
-        # Multiple ff groups, if cohort properties are OR'ed together.
-        from posthog.models.property.util import clear_excess_levels
-
-        target_properties = clear_excess_levels(cohort.properties)
-
-        if isinstance(target_properties, Property):
-            # cohort was effectively a single property.
-            parsed_conditions.append(
-                {
-                    "properties": [target_properties.to_dict()],
-                    "rollout_percentage": cohort_group_rollout,
-                }
-            )
-
-        elif isinstance(target_properties.values[0], Property):
-            # Property Group of properties
-            if target_properties.type == PropertyOperatorType.AND:
-                parsed_conditions.append(
-                    {
-                        "properties": [prop.to_dict() for prop in target_properties.values],
-                        "rollout_percentage": cohort_group_rollout,
-                    }
-                )
-            else:
-                # cohort OR requires multiple ff group
-                for prop in target_properties.values:
-                    parsed_conditions.append(
-                        {
-                            "properties": [prop.to_dict()],
-                            "rollout_percentage": cohort_group_rollout,
-                        }
-                    )
-        else:
-            # If there's nested property groups, we need to express that as OR of ANDs.
-            # Being a bit dumb here, and not trying to apply De Morgan's law to coerce AND of ORs into OR of ANDs.
-            if target_properties.type == PropertyOperatorType.AND:
-                return self.conditions
-
-            for prop_group in cast(List[PropertyGroup], target_properties.values):
-                if (
-                    len(prop_group.values) == 0
-                    or not isinstance(prop_group.values[0], Property)
-                    or (prop_group.type == PropertyOperatorType.OR and len(prop_group.values) > 1)
-                ):
-                    # too nested or invalid, bail out
-                    return self.conditions
-
-                parsed_conditions.append(
-                    {
-                        "properties": [prop.to_dict() for prop in prop_group.values],
-                        "rollout_percentage": cohort_group_rollout,
-                    }
-                )
-
-        return parsed_conditions
-
-    @property
-    def cohort_ids(self) -> List[int]:
-        cohort_ids = []
-        for condition in self.conditions:
-            props = condition.get("properties", [])
-            for prop in props:
-                if prop.get("type") == "cohort":
-                    cohort_id = prop.get("value")
-                    if cohort_id:
-                        cohort_ids.append(cohort_id)
-        return cohort_ids
-
-    def update_cohorts(self) -> None:
-        from posthog.tasks.calculate_cohort import update_cohort
-        from posthog.tasks.cohorts_in_feature_flag import COHORT_ID_IN_FF_KEY
-
-        if self.cohort_ids:
-            cache.delete(COHORT_ID_IN_FF_KEY)
-            for cohort in Cohort.objects.filter(pk__in=self.cohort_ids):
-                update_cohort(cohort)
-
-    def __str__(self):
-        return f"{self.key} ({self.pk})"
-
-
-@mutable_receiver(pre_delete, sender=Experiment)
-def delete_experiment_flags(sender, instance, **kwargs):
-    FeatureFlag.objects.filter(experiment=instance).update(deleted=True)
-
-
-class FeatureFlagHashKeyOverride(models.Model):
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["team", "person", "feature_flag_key"], name="Unique hash_key for a user/team/feature_flag combo"
-            )
-        ]
-
-    # Can't use a foreign key to feature_flag_key directly, since
-    # the unique constraint is on (team_id+key), and not just key.
-    # A standard id foreign key leads to INNER JOINs everytime we want to get the key
-    # and we only ever want to get the key.
-    feature_flag_key: models.CharField = models.CharField(max_length=400)
-    person: models.ForeignKey = models.ForeignKey("Person", on_delete=models.CASCADE)
-    team: models.ForeignKey = models.ForeignKey("Team", on_delete=models.CASCADE)
-    hash_key: models.CharField = models.CharField(max_length=400)
-
+from .feature_flag import (
+    FeatureFlag,
+    FeatureFlagHashKeyOverride,
+    get_feature_flags_for_team_in_cache,
+    set_feature_flags_for_team_in_cache,
+)
 
 __LONG_SCALE__ = float(0xFFFFFFFFFFFFFFF)
 
@@ -600,41 +364,6 @@ class FeatureFlagMatcher:
         return current_match, current_index
 
 
-def set_feature_flags_for_team_in_cache(
-    team_id: int, feature_flags: Optional[List[FeatureFlag]] = None
-) -> List[FeatureFlag]:
-    from posthog.api.feature_flag import MinimalFeatureFlagSerializer
-
-    if feature_flags is not None:
-        all_feature_flags = feature_flags
-    else:
-        all_feature_flags = list(FeatureFlag.objects.filter(team_id=team_id, active=True, deleted=False))
-
-    serialized_flags = MinimalFeatureFlagSerializer(all_feature_flags, many=True).data
-
-    cache.set(f"team_feature_flags_{team_id}", json.dumps(serialized_flags), None)
-
-    return all_feature_flags
-
-
-def get_feature_flags_for_team_in_cache(team_id: int) -> Optional[List[FeatureFlag]]:
-    try:
-        flag_data = cache.get(f"team_feature_flags_{team_id}")
-    except Exception:
-        # redis is unavailable
-        return None
-
-    if flag_data is not None:
-        try:
-            parsed_data = json.loads(flag_data)
-            return [FeatureFlag(**flag) for flag in parsed_data]
-        except Exception as e:
-            capture_exception(e)
-            return None
-
-    return None
-
-
 def hash_key_overrides(team_id: int, person_id: int) -> Dict[str, str]:
     feature_flag_to_key_overrides = {}
     for feature_flag, override in FeatureFlagHashKeyOverride.objects.filter(
@@ -788,139 +517,3 @@ def set_feature_flag_hash_key_overrides(
         # at the same time. In this case, we can safely ignore the error.
         # We don't want to return an error response for `/decide` just because of this.
         FeatureFlagHashKeyOverride.objects.bulk_create(new_overrides, ignore_conflicts=True)
-
-
-def get_user_blast_radius(team: Team, feature_flag_condition: dict, group_type_index: Optional[GroupTypeIndex] = None):
-
-    from posthog.queries.person_query import PersonQuery
-
-    # No rollout % calculations here, since it makes more sense to compute that on the frontend
-    properties = feature_flag_condition.get("properties") or []
-
-    if group_type_index is not None:
-
-        try:
-            from ee.clickhouse.queries.groups_join_query import GroupsJoinQuery
-        except Exception:
-            return 0, 0
-
-        if len(properties) > 0:
-            filter = Filter(data=feature_flag_condition, team=team)
-
-            for property in filter.property_groups.flat:
-                if property.group_type_index is None or (property.group_type_index != group_type_index):
-                    raise ValidationError("Invalid group type index for feature flag condition.")
-
-            groups_query, groups_query_params = GroupsJoinQuery(filter, team.id).get_filter_query(
-                group_type_index=group_type_index
-            )
-
-            total_affected_count = sync_execute(
-                f"""
-                SELECT count(1) FROM (
-                    {groups_query}
-                )
-            """,
-                groups_query_params,
-            )[0][0]
-        else:
-            total_affected_count = team.groups_seen_so_far(group_type_index)
-
-        return total_affected_count, team.groups_seen_so_far(group_type_index)
-
-    if len(properties) > 0:
-        filter = Filter(data=feature_flag_condition, team=team)
-        cohort_filters = []
-        for property in filter.property_groups.flat:
-            if property.type in ["cohort", "precalculated-cohort", "static-cohort"]:
-                cohort_filters.append(property)
-
-        target_cohort = None
-
-        if len(cohort_filters) == 1:
-            try:
-                target_cohort = Cohort.objects.get(id=cohort_filters[0].value, team=team)
-            except Cohort.DoesNotExist:
-                pass
-            finally:
-                cohort_filters = []
-
-        person_query, person_query_params = PersonQuery(
-            filter, team.id, cohort=target_cohort, cohort_filters=cohort_filters
-        ).get_query()
-
-        total_count = sync_execute(
-            f"""
-            SELECT count(1) FROM (
-                {person_query}
-            )
-        """,
-            person_query_params,
-        )[0][0]
-
-    else:
-        total_count = team.persons_seen_so_far
-
-    blast_radius = total_count
-    total_users = team.persons_seen_so_far
-
-    return blast_radius, total_users
-
-
-def can_user_edit_feature_flag(request, feature_flag):
-    # self hosted check for enterprise models that may not exist
-    try:
-        from ee.models.feature_flag_role_access import FeatureFlagRoleAccess
-        from ee.models.organization_resource_access import OrganizationResourceAccess
-    except:
-        return True
-    else:
-        if not request.user.organization.is_feature_available(AvailableFeature.ROLE_BASED_ACCESS):
-            return True
-        if feature_flag.created_by == request.user:
-            return True
-        if (
-            request.user.organization_memberships.get(organization=request.user.organization).level
-            >= OrganizationMembership.Level.ADMIN
-        ):
-            return True
-        all_role_memberships = request.user.role_memberships.select_related("role").all()
-        try:
-            feature_flag_resource_access = OrganizationResourceAccess.objects.get(
-                organization=request.user.organization, resource=OrganizationResourceAccess.Resources.FEATURE_FLAGS
-            )
-            if feature_flag_resource_access.access_level >= OrganizationResourceAccess.AccessLevel.CAN_ALWAYS_EDIT:
-                return True
-            org_level = feature_flag_resource_access.access_level
-        except OrganizationResourceAccess.DoesNotExist:
-            org_level = OrganizationResourceAccess.AccessLevel.CAN_ALWAYS_EDIT
-
-        role_level = max([membership.role.feature_flags_access_level for membership in all_role_memberships], default=0)
-
-        if role_level == 0:
-            final_level = org_level
-        else:
-            final_level = role_level
-        if final_level == OrganizationResourceAccess.AccessLevel.CAN_ONLY_VIEW:
-            can_edit = FeatureFlagRoleAccess.objects.filter(
-                feature_flag__id=feature_flag.pk,
-                role__id__in=[membership.role.pk for membership in all_role_memberships],
-            ).exists()
-            return can_edit
-        else:
-            return final_level == OrganizationResourceAccess.AccessLevel.CAN_ALWAYS_EDIT
-
-
-# DEPRECATED: This model is no longer used, but it's not deleted to avoid downtime
-class FeatureFlagOverride(models.Model):
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["user", "feature_flag", "team"], name="unique feature flag for a user/team combo"
-            )
-        ]
-
-    feature_flag: models.ForeignKey = models.ForeignKey("FeatureFlag", on_delete=models.CASCADE)
-    user: models.ForeignKey = models.ForeignKey("User", on_delete=models.CASCADE)
-    override_value: models.JSONField = models.JSONField()
-    team: models.ForeignKey = models.ForeignKey("Team", on_delete=models.CASCADE)
